Create MSC dissertation
This folder includes the app and the report of my final project during my MSc studies. It is an interactive map of Portsmouth tracing Sir Arthur Conan Doyle's movements in the city. Sir Arthur Conan Doyle (the writer of Sherlock Holmes series) lived and worked in Portsmouth for 8 years. The project was to create a mobile application that presents a guided tour to sites of interest around Portsmouth to visit locations related to Conan Doyle and other notable events.